Calculate winAscent and descent in all assemble scripts
