+++ scripts/assembleTools.py
@@ -4,6 +4,20 @@ from fontTools.pens.reverseContourPen import ReverseContourPen
 from fontTools.pens.transformPen import TransformPointPen
 
 
+def computeWinAscentDescent(font):
+    fontYMin = 0
+    fontYMax = 0
+    for glyph in font:
+        bounds = glyph.getBounds(font)
+        if bounds is None:
+            continue  # empty glyph
+        xMin, yMin, xMax, yMax = bounds
+        fontYMin = min(fontYMin, yMin)
+        fontYMax = max(fontYMax, yMax)
+    font.info.openTypeOS2WinAscent = fontYMax
+    font.info.openTypeOS2WinDescent = abs(fontYMin)
+
+
 class DecomposingRecordingPointPen(RecordingPointPen):
     def __init__(self, glyphSet):
         super(DecomposingRecordingPointPen, self).__init__()
